--- src/main/java/KirbyException.java
@@ -1,10 +1,0 @@
-/**
- * The KirbyException class implements the exception class specific
- * to this program.
- * @author Sheryl-Lynn Tan (G11)
- */
-public class KirbyException extends Exception {
-    KirbyException(String message) {
-        super(message);
-    }
-}
