+++ src/main/java/kirby/exceptions/KirbyException.java
@@ -1,0 +1,12 @@
+package kirby.exceptions;
+
+/**
+ * The kirby.KirbyException class implements the exception class specific
+ * to this program.
+ * @author Sheryl-Lynn Tan (G11)
+ */
+public class KirbyException extends Exception {
+    public KirbyException(String message) {
+        super(message);
+    }
+}
