+++ src/main/java/kirby/exceptions/KirbyInvalidCommandException.java
@@ -1,10 +1,12 @@
+package kirby.exceptions;
+
 /**
- * The KirbyInvalidCommandException class inherits from KirbyException
+ * The kirby.KirbyInvalidCommandException class inherits from kirby.KirbyException
  * and is thrown when there is an undefined input command.
  * @author Sheryl-Lynn Tan (G11)
  */
-public class KirbyInvalidCommandException extends KirbyException{
-    KirbyInvalidCommandException() {
+public class KirbyInvalidCommandException extends KirbyException {
+    public KirbyInvalidCommandException() {
         super("Aaaaaaah, I don't understand you :(\n" +
                 "Try adding a valid command (todo, deadline, event)");
     }
