+++ src/main/java/kirby/exceptions/KirbyMissingArgumentException.java
@@ -1,10 +1,12 @@
+package kirby.exceptions;
+
 /**
- * The KirbyMissingArgumentException class inherits from KirbyException
+ * The kirby.KirbyMissingArgumentException class inherits from kirby.KirbyException
  * and is thrown when there is an invalid input command argument.
  * @author Sheryl-Lynn Tan (G11)
  */
 public class KirbyMissingArgumentException extends KirbyException {
-    KirbyMissingArgumentException(String message) {
+    public KirbyMissingArgumentException(String message) {
         super("Aaaaaaah, the description of a " + message + " is invalid! Try again!" );
     }
 }
